+++ templates/workspace_services/guacamole/guacamole-server/guacamole-auth-azure/src/main/java/org/apache/guacamole/auth/azuretre/connection/ConnectionService.java
@@ -111,6 +111,11 @@ public class ConnectionService {
         config.setParameter("drive-path", System.getenv("GUAC_DRIVE_PATH"));
         config.setParameter("disable-download", System.getenv("GUAC_DISABLE_DOWNLOAD"));
         config.setParameter("disable-upload", System.getenv("GUAC_DISABLE_UPLOAD"));
+
+        String serverLayout = System.getenv("GUAC_SERVER_LAYOUT");
+        if (serverLayout != null)
+          if (!serverLayout.isEmpty())
+            config.setParameter("server-layout", serverLayout);
     }
 
     private static JSONArray getVMsFromProjectAPI(final AzureTREAuthenticatedUser user) throws GuacamoleException {